(IT Case) Refactor the FileSinkITCase to reuse the sink components and checking logic
