+++ flink-connectors/flink-connector-files/src/test/java/org/apache/flink/connector/file/sink/FileSinkITBase.java
@@ -18,17 +18,14 @@
 
 package org.apache.flink.connector.file.sink;
 
-import org.apache.flink.api.common.serialization.Encoder;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.RestOptions;
+import org.apache.flink.connector.file.sink.utils.FileSinkIntegerTestDataUtils;
 import org.apache.flink.core.fs.Path;
-import org.apache.flink.core.io.SimpleVersionedSerializer;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.minicluster.MiniCluster;
 import org.apache.flink.runtime.minicluster.MiniClusterConfiguration;
-import org.apache.flink.streaming.api.functions.sink.filesystem.BucketAssigner;
 import org.apache.flink.streaming.api.functions.sink.filesystem.PartFileInfo;
-import org.apache.flink.streaming.api.functions.sink.filesystem.bucketassigners.SimpleVersionedStringSerializer;
 import org.apache.flink.streaming.api.functions.sink.filesystem.rollingpolicies.CheckpointRollingPolicy;
 import org.apache.flink.util.TestLogger;
 
@@ -37,22 +34,9 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runners.Parameterized;
 
-import java.io.DataInputStream;
-import java.io.EOFException;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.Map;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 
 /**
  * The base class for the File Sink IT Case in different execution mode.
@@ -101,86 +85,17 @@ public abstract class FileSinkITBase extends TestLogger {
 			miniCluster.executeJobBlocking(jobGraph);
 		}
 
-		checkResult(path);
+		FileSinkIntegerTestDataUtils.checkIntegerSequenceSinkOutput(path, NUM_RECORDS, NUM_BUCKETS, NUM_SOURCES);
 	}
 
 	protected abstract JobGraph createJobGraph(String path);
 
 	protected FileSink<Integer> createFileSink(String path) {
 		return FileSink
-				.forRowFormat(new Path(path), new IntEncoder())
-				.withBucketAssigner(new ModuloBucketAssigner())
+				.forRowFormat(new Path(path), new FileSinkIntegerTestDataUtils.IntEncoder())
+				.withBucketAssigner(new FileSinkIntegerTestDataUtils.ModuloBucketAssigner(NUM_BUCKETS))
 				.withRollingPolicy(new PartSizeAndCheckpointRollingPolicy(1024))
 				.build();
-	}
-
-	private void checkResult(String path) throws Exception {
-		File dir = new File(path);
-		String[] subDirNames = dir.list();
-		assertNotNull(subDirNames);
-
-		Arrays.sort(subDirNames, Comparator.comparingInt(Integer::parseInt));
-		assertEquals(NUM_BUCKETS, subDirNames.length);
-		for (int i = 0; i < NUM_BUCKETS; ++i) {
-			assertEquals(Integer.toString(i), subDirNames[i]);
-
-			// now check its content
-			File bucketDir = new File(path, subDirNames[i]);
-			assertTrue(
-					bucketDir.getAbsolutePath() + " Should be a existing directory",
-					bucketDir.isDirectory());
-
-			Map<Integer, Integer> counts = new HashMap<>();
-			File[] files = bucketDir.listFiles(f -> !f.getName().startsWith("."));
-			assertNotNull(files);
-
-			for (File file : files) {
-				assertTrue(file.isFile());
-
-				try (DataInputStream dataInputStream = new DataInputStream(new FileInputStream(file))) {
-					while (true) {
-						int value = dataInputStream.readInt();
-						counts.compute(value, (k, v) -> v == null ? 1 : v + 1);
-					}
-				} catch (EOFException e) {
-					// End the reading
-				}
-			}
-
-			int expectedCount = NUM_RECORDS / NUM_BUCKETS +
-					(i < NUM_RECORDS % NUM_BUCKETS ? 1 : 0);
-			assertEquals(expectedCount, counts.size());
-
-			for (int j = i; j < NUM_RECORDS; j += NUM_BUCKETS) {
-				assertEquals(
-						"The record " + j + " should occur " + NUM_SOURCES + " times, " +
-								" but only occurs " + counts.getOrDefault(j, 0) + "time",
-						NUM_SOURCES,
-						counts.getOrDefault(j, 0).intValue());
-			}
-		}
-	}
-
-	private static class IntEncoder implements Encoder<Integer> {
-
-		@Override
-		public void encode(Integer element, OutputStream stream) throws IOException {
-			stream.write(ByteBuffer.allocate(4).putInt(element).array());
-			stream.flush();
-		}
-	}
-
-	private static class ModuloBucketAssigner implements BucketAssigner<Integer, String> {
-
-		@Override
-		public String getBucketId(Integer element, Context context) {
-			return Integer.toString(element % NUM_BUCKETS);
-		}
-
-		@Override
-		public SimpleVersionedSerializer<String> getSerializer() {
-			return SimpleVersionedStringSerializer.INSTANCE;
-		}
 	}
 
 	private static class PartSizeAndCheckpointRollingPolicy
